Change the version to 0.3.7
This is a horrible thing to do, because only the first party should
publish stable builds of their product.

However, I don't really have much choice, because nirs generated by
prereleases are incompatible with nirs generated by other releases.
This means that using a prerelease precludes using already published
artifacts built with other releases.

I don't fancy rebuilding and publishing the entire ScalaPB runtime,
so I decided to do the horrible thing.

To avoid the potential problems, I'll be using the staging repository
provided by Sonatype: https://oss.sonatype.org/content/repositories/staging.
Everything that is published to Sonatype, closed but not yet released,
ends up in this repository.

This means that all the Scala Native "0.3.7" artifacts that I'll be publishing
are going to be temporary - they will be good enough for the Scalameta build,
and we'll be able to drop them when the official 0.3.7 release comes out.